+++ test/test_sendgrid.py
@@ -2308,7 +2308,7 @@ class UnitTests(unittest.TestCase):
         with open(LICENSE_FILE, 'r') as f:
             copyright_line = f.readline().rstrip()
         self.assertEqual(
-            'Copyright (c) 2012-%s SendGrid, Inc.' % datetime.datetime.now().year,
+            'Copyright (c) 2012-%s Twilio SendGrid, Inc.' % datetime.datetime.now().year,
             copyright_line)
 
     # @classmethod